+++ src/components/human-labelling/AnnotationJobView.tsx
@@ -31,6 +31,9 @@ type Job = {
   status: "pending" | "in_progress" | "completed";
   created_at: string;
   completed_at: string | null;
+  /** Backend marks the read-only viewer link state on completed jobs. */
+  is_public?: boolean;
+  view_token?: string | null;
 };
 
 type Annotator = { uuid: string; name: string };
@@ -81,6 +84,8 @@ type JobResponse = {
   evaluators: Evaluator[];
   items: Item[];
   annotations: Annotation[];
+  /** True when fetched via the read-only viewer token; false on the annotator route. */
+  read_only?: boolean;
 };
 
 type LoadState =
@@ -143,13 +148,20 @@ async function publicFetch<T>(
   return { ok: true, data };
 }
 
-export type AnnotationJobMode = "public" | "admin";
+export type AnnotationJobMode = "public" | "admin" | "public-readonly";
 
 export type AnnotationJobMeta = {
   task: { uuid: string; name: string; type: string };
   annotator: { uuid: string; name: string };
   jobStatus: "pending" | "in_progress" | "completed";
   evaluators: { uuid: string; name: string }[];
+  /** Job UUID + share state, threaded through so the admin wrapper can render
+   * the visibility toggle for the read-only viewer link. */
+  job: {
+    uuid: string;
+    is_public: boolean;
+    view_token: string | null;
+  };
 };
 
 export function AnnotationJobView({
@@ -172,6 +184,8 @@ export function AnnotationJobView({
   const [submitting, setSubmitting] = useState(false);
   const [topError, setTopError] = useState<string | null>(null);
 
+  const isReadOnlyMode = mode === "admin" || mode === "public-readonly";
+
   const initialise = useCallback(
     (data: JobResponse) => {
       const next: Record<FieldKey, FieldValue> = {};
@@ -188,11 +202,11 @@ export function AnnotationJobView({
       setFields(next);
       setSavedKeys(saved);
 
-      // Admin (read-only) view always starts on the first item — admins
-      // are reviewing what's been labelled, not picking up where the
-      // annotator left off. Write mode jumps to the first item that
+      // Read-only views (admin, public-readonly) always start on the first
+      // item — they're reviewing what's been labelled, not picking up where
+      // the annotator left off. Write mode jumps to the first item that
       // still has at least one unlabelled evaluator.
-      if (mode === "admin") {
+      if (isReadOnlyMode) {
         setCurrentIndex(0);
         return;
       }
@@ -201,16 +215,20 @@ export function AnnotationJobView({
       );
       setCurrentIndex(firstIncomplete >= 0 ? firstIncomplete : 0);
     },
-    [mode],
+    [isReadOnlyMode],
   );
 
   useEffect(() => {
     if (!token) return;
     let cancelled = false;
+    // The viewer-token endpoint is mounted under …/jobs/view/{token}; the
+    // annotator's own (read+write) endpoint stays at …/jobs/{token}.
+    const fetchPath =
+      mode === "public-readonly"
+        ? `/public/annotation-jobs/view/${encodeURIComponent(token)}`
+        : `/public/annotation-jobs/${encodeURIComponent(token)}`;
     const run = async () => {
-      const result = await publicFetch<JobResponse>(
-        `/public/annotation-jobs/${encodeURIComponent(token)}`,
-      );
+      const result = await publicFetch<JobResponse>(fetchPath);
       if (cancelled) return;
       if (!result.ok) {
         if (result.status === 404) setState({ status: "not_found" });
@@ -238,6 +256,11 @@ export function AnnotationJobView({
           uuid: e.uuid,
           name: e.name,
         })),
+        job: {
+          uuid: result.data.job.uuid,
+          is_public: result.data.job.is_public ?? false,
+          view_token: result.data.job.view_token ?? null,
+        },
       });
     };
     run().catch((err) => {
@@ -307,7 +330,9 @@ export function AnnotationJobView({
   }
 
   const { data } = state;
-  const isAdmin = mode === "admin";
+  // Inner view treats `isAdmin` as a generic read-only flag — both the admin
+  // wrapper and the public read-only viewer should disable writes.
+  const isAdmin = isReadOnlyMode;
   const items = data.items;
   const evaluators = data.evaluators;
   const total = items.length;
